services/visualization/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.1
jinja2==3.1.2
pyyaml==6.0.1
graphviz==0.20.1
matplotlib==3.8.2
plotly==5.18.0
networkx==3.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
